+++ src/routes/roomRoutes/private/routes.js
@@ -6,6 +6,7 @@ import ApplicationError from '../../../errors/ApplicationError';
 
 const router = Router();
 
+
 router.get('/list', async (req, res, next) => {
     try {
   
@@ -19,11 +20,17 @@ router.get('/list', async (req, res, next) => {
 
 router.post('/create', async (req, res, next) => {
     try {
-      const newRoomInfo = req.body;
+      const role = req.user.role
+      if (role === "síndico") {
+        const newRoomInfo = req.body;
   
-      await RoomsService.create(newRoomInfo);
+        await RoomsService.create(newRoomInfo);
   
-      return res.status(201).json();
+        return res.status(201).json();
+      } else{
+        return res.status(403).json({message:"Access Denied"});
+      }
+      
     } catch (error) {
       return next(new ApplicationError(error));
     }
@@ -31,12 +38,18 @@ router.post('/create', async (req, res, next) => {
 
 router.put('/update/:id',async (req, res, next) => {
     try {
-      const { id } = req.params;
-      const updateObject = req.body;
+      const role = req.user.role
+      if (role === "síndico") {
+        const { id } = req.params;
+        const updateObject = req.body;
   
-      const updatedRoom = await projectsService.updateOne(updateObject, id);
+        const updatedRoom = await projectsService.updateOne(updateObject, id);
   
-      return res.status(200).json(updatedRoom);
+        return res.status(200).json(updatedRoom);
+      }else{
+        return res.status(403).json({message:"Access Denied"});
+      }
+
     } catch (error) {
       return next(new ApplicationError(error));
     }
@@ -44,11 +57,17 @@ router.put('/update/:id',async (req, res, next) => {
 
 router.delete('/delete/:id', async (req, res, next) => {
     try {
-      const { id } = req.params;
+      const role = req.user.role
+      if (role === "síndico") {
+        const { id } = req.params;
   
-      await RoomsService.deleteOne(id);
+        await RoomsService.deleteOne(id);
   
-      return res.status(200).json();
+        return res.status(200).json();
+      }else{
+        return res.status(403).json({message:"Access Denied"});
+      }
+
     } catch (error) {
       return next(new ApplicationError(error));
     }
